refactor: re-inline single-consumer services per ADR-0012
Re-inline 9 single-consumer services back into handler files (Tier 1).
Relocate 3 services to handler-local sibling files (Tier 2):
- ApiGatewayAuthorizer/helpers.ts (authorizerService + keyService)
- SendPushNotification/push-helpers.ts
- SendPushNotification/endpoint-cleanup-helpers.ts

12 service files deleted, 4 empty directories removed.
All handlers under 150 lines. Zero mantle check violations.

Confidence: high
Scope-risk: moderate
