+++ src/lambdas/sqs/SendPushNotification/endpoint-cleanup-helpers.ts
